backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
pydantic-settings==2.5.0
pydantic[email]==2.8.2
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
cryptography==43.0.0
telethon==1.37.0
aiogram==3.13.0
redis==5.1.0
slowapi==0.1.9
python-multipart==0.0.12
